refactor: move color constants to a dedicated file for better organization
