+++ apps/web/src/constants/idb.ts
@@ -1,4 +1,4 @@
-import { color } from "#/DB/color";
+import { color } from "#/constants/color";
 import { ParticipantIDBData } from "#/types/idb";
 
 export enum IDBStore {
