+++ apps/web/src/routes/kiwi/-split-view/annotation/tabs/highlight/index.tsx
@@ -3,7 +3,7 @@ import { useAtomValue } from "@bookiwi/jotai";
 import Annotation from "./annotation";
 import NotSelected from "./not-selected";
 
-import { primaryColor } from "#/DB/color";
+import { primaryColor } from "#/constants/color";
 import { selectedAnnotationAtom } from "#/routes/kiwi/-reader/atoms/annotations";
 import { participantColorAtom } from "#/routes/kiwi/-reader/atoms/participants";
