+++ apps/web/src/routes/my-kiwis/-kiwis/create-kiwi-modal/index.tsx
@@ -14,7 +14,6 @@ import KiwiInfo from "./kiwi-info";
 import Loading from "./loading";
 import { ActionTypes, Step } from "./reducer";
 
-import { color } from "#/DB/color";
 import tempUser from "#/DB/users";
 import {
   Dialog,
@@ -24,6 +23,7 @@ import {
   DialogTitle,
   DialogFooter,
 } from "#/components/ui/dialog";
+import { color } from "#/constants/color";
 import { IDBStore } from "#/constants/idb";
 import idb from "#/managers/idb";
 import { useKiwis } from "#/routes/my-kiwis/-context";
